requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
django-mathfilters==1.0.0
lxml==4.9.1
Pillow==9.3.0
python-docx==0.8.11
sqlparse==0.4.3
